Added SearchBar as own componen and implemented simple search-function for filtering result of searched movies....in progress
